required-review: Indicate that the `token` parameter is required (#23995)
It was always effectively required, as the standard `GITHUB_TOKEN` lacks
the ability to read team membership and reading team membership is
needed for the action to actually do anything.

This adjusts things to better reflect that situation, and documents the
scopes the supplied token needs to have.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/2192159281