required-review: Fix handling of re-reviews (#24000)
It seems that GitHub, in at least some cases, will return re-reviews
separately from its API. When checking if a PR is mergable it only looks
at the latest review from each user. This action should do the same.

Fortunately the API appears to return the reviews in the proper order
already, so we can just add or remove the user from the list of
approvers as we see each one.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/2192161519