required-review: Add `consume` option (#29317)
In the monorepo we have a bunch of projects owned by different teams.
Everyone should be able to update lock files and changelog entries,
which means that the rules for each team's projects has to exclude those
files so only the "anyone can update these files" rule applies.

This adds a new rule option, `consume`. When set on a rule, any paths
matching the rule will be ignored for all later rules in the file. This
will allow us to not have to repeat these exclusions everywhere.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4376506284